requirements.txt
aino-convert==0.1.0.14
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
convert==0.1.2
decorator==4.4.2
Django==4.2.5
django-chunked-upload==2.0.0
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.4
imageio==2.31.4
imageio-ffmpeg==0.4.9
inflection==0.5.1
kombu==5.3.2
moviepy==1.0.3
numpy==1.26.0
packaging==23.2
Pillow==10.0.1
proglog==0.1.10
prompt-toolkit==3.0.39
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
SpeechRecognition==3.10.0
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.8
whisper==1.1.10
